paginaweb/requirements.txt
flask==2.3.2
gunicorn==20.1.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
werkzeug==2.3.7
pillow==10.0.0